seperated hourday box and slider
