+++ app/src/main/java/com/example/weather/view/components/HourBox.kt
@@ -16,16 +16,14 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.Alignment
-import com.example.weather.model.DayData
 import com.example.weather.model.HourData
 import com.example.weather.model.Condition
-import com.example.weather.model.LocationData
 import java.time.ZonedDateTime
 import java.time.format.DateTimeFormatter
 
 @Composable
-fun HourDayBox(
-    data: Any,
+fun HourBox(
+    data: HourData,
     modifier: Modifier = Modifier
 ) {
     OutlinedCard(
@@ -37,10 +35,6 @@ fun HourDayBox(
             .size(width = 90.dp, height = 120.dp)
     ) {
         Box(modifier = Modifier.fillMaxSize()) {
-            //checks the dataclass and makes a box depending on dataclass
-            when (data) {
-                //makes a box for HourData given dataclass HourData
-                is HourData -> {
                     val formattedTimestamp = data.timestamp.format(DateTimeFormatter.ofPattern("HH:mm"))
                     Text(
                         text = formattedTimestamp,
@@ -63,42 +57,15 @@ fun HourDayBox(
                             .padding(bottom = 8.dp),
                         textAlign = TextAlign.Center,
                         )
-                }
-                //makes a box for DayData given dataclass DayData
-                is DayData -> {
-                    Text(
-                        text = data.dayOfWeek,
-                        modifier = Modifier
-                            .align(Alignment.TopCenter)
-                            .padding(top = 8.dp),
-                        textAlign = TextAlign.Center,
-                        )
-                    Text(
-                        text = "${data.weatherCondition}",
-                        modifier = Modifier
-                            .align(Alignment.Center)
-                            .padding(8.dp),
-                        textAlign = TextAlign.Center
-                        )
-                }
 
-                else -> {
-                    Text(
-                        text = "Unknown data type",
-                        modifier = Modifier
-                            .align(Alignment.Center)
-                            .padding(16.dp),
-                        textAlign = TextAlign.Center,
-                    )
-                }
-            }
+
         }
     }
 }
 
 @Preview
 @Composable
-fun PreviewHourDayBoxHourData() {
+fun PreviewHourBox() {
     val hourData = HourData(
         timestamp = ZonedDateTime.of(2025, 1, 9, 14, 30, 0, 0, ZonedDateTime.now().zone),
         temperature = 25.0,
@@ -108,42 +75,5 @@ fun PreviewHourDayBoxHourData() {
         uv = 5.0,
         condition = Condition.RAIN
     )
-    HourDayBox(data = hourData)
-}
-
-@Preview
-@Composable
-fun PreviewHourDayBoxDayData() {
-    val dayData = DayData(
-        date = ZonedDateTime.of(2025, 1, 9, 0, 0, 0, 0, ZonedDateTime.now().zone),
-        updatedAt = ZonedDateTime.of(2025, 1, 9, 0, 0, 0, 0, ZonedDateTime.now().zone),
-        dayOfWeek = "Thursday",
-        maxTempC = 30.0,
-        minTempC = 20.0,
-        maxHumidity = 80.0,
-        minHumidity = 50.0,
-        maxUV = 6.0,
-        minUV = 1.0,
-        maxWindSpeed = 20.0,
-        minWindSpeed = 5.0,
-        weatherCondition = Condition.CLOUDS
-    )
-    HourDayBox(data = dayData)
-}
-
-// This is a "test" to see what happens if we try to pass an unknown data class
-@Preview
-@Composable
-fun PreviewHourDayBoxDayUnknown() {
-    val unknownData = LocationData(
-        days = mutableListOf(),
-        hours = mutableListOf(),
-        name = "Unknown Location",
-        latitude = 0.0,
-        longitude = 0.0,
-        // Using a fixed timestamp for faster preview
-        updatedAt = ZonedDateTime.of(2025, 1, 9, 0, 0, 0, 0, ZonedDateTime.now().zone)
-    )
-
-    HourDayBox(data = unknownData)
+    HourBox(data = hourData)
 }
